refactor people related classes
